# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "l-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/nodes/get/{id}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Fetch a single node by ID.",
        "operationId": "get_panel_api_nodes_get_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/webCert/{id}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Fetch a node's own web TLS certificate/key file paths (proxied to the node). Used by the inbound form's \"Set Cert from Panel\" so a node-assigned inbound gets paths that exist on the node, not the central panel.",
        "operationId": "get_panel_api_nodes_webCert_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "webCertFile": "/root/cert/example.com/fullchain.pem",
                    "webKeyFile": "/root/cert/example.com/privkey.pem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/add": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Register a new remote node. Provide its URL, apiToken, and optional remark / allowPrivateAddress flag.",
        "operationId": "post_panel_api_nodes_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "remark": "",
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef...",
                "enable": true,
                "allowPrivateAddress": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/bootstrap": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "SSH into a new host, install the l-ui agent, start the service, and register the node in one step. The panel generates the API token and returns the executed bootstrap steps for display in the UI.",
        "operationId": "post_panel_api_nodes_bootstrap",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "remark": "",
                "address": "203.0.113.10",
                "sshUser": "root",
                "sshPassword": "secret",
                "sshPort": 22,
                "agentPort": 2053
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "node": {
                      "id": 1,
                      "name": "de-fra-1",
                      "address": "203.0.113.10",
                      "port": 2053,
                      "basePath": "/",
                      "enable": true
                    },
                    "steps": [
                      {
                        "name": "detect-arch",
                        "ok": true,
                        "output": "x86_64"
                      },
                      {
                        "name": "verify-agent",
                        "ok": true,
                        "output": "{...}"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/update/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Replace a node’s connection details. Same body shape as /add.",
        "operationId": "post_panel_api_nodes_update_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "de-fra-1",
                "remark": "",
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef...",
                "enable": true,
                "allowPrivateAddress": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/del/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Delete a node. Inbounds bound to it are not auto-migrated.",
        "operationId": "post_panel_api_nodes_del_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/setEnable/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Pause or resume traffic sync with this node.",
        "operationId": "post_panel_api_nodes_setEnable_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "enable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/test": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Probe a node without saving it. Uses the body as connection details and returns the same heartbeat snapshot a registered node would have.",
        "operationId": "post_panel_api_nodes_test",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/",
                "apiToken": "abcdef..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "status": "online",
                    "latencyMs": 42,
                    "xrayVersion": "25.x.x",
                    "panelVersion": "v3.x.x",
                    "cpuPct": 12.5,
                    "memPct": 45.2,
                    "uptimeSecs": 86400,
                    "error": ""
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/certFingerprint": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Connect to the node over HTTPS without verifying its certificate and return the leaf certificate's SHA-256 (base64). Used by the Add/Edit Node dialog to fetch and pin a self-signed certificate. Uses the same body as /test.",
        "operationId": "post_panel_api_nodes_certFingerprint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "scheme": "https",
                "address": "node1.example.com",
                "port": 2053,
                "basePath": "/"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": "k3b1...base64-sha256...="
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/probe/{id}": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Probe an existing node, updating its cached health state.",
        "operationId": "post_panel_api_nodes_probe_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/updatePanel": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Trigger the official panel self-updater on each given node (downloads the latest release and restarts). Only enabled, online nodes are updated; offline/disabled ones are reported as skipped. Returns a per-node result list.",
        "operationId": "post_panel_api_nodes_updatePanel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "ids": [
                  1,
                  2,
                  3
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": [
                    {
                      "id": 1,
                      "name": "de-1",
                      "ok": true
                    },
                    {
                      "id": 2,
                      "name": "fr-1",
                      "ok": false,
                      "error": "node is offline"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/nodes/history/{id}/{metric}/{bucket}": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Aggregated metric history for a node — same shape as /server/history, scoped to one node.",
        "operationId": "get_panel_api_nodes_history_id_metric_bucket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Node ID.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "metric",
            "in": "path",
            "required": true,
            "description": "cpu | mem.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "description": "Bucket size in seconds. Allowed: 2, 30, 60, 120, 180, 300.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}